\name{Classifier}
\Rdversion{1.1}
\docType{class}
\alias{Classifier}
\alias{Classifier-class}
\alias{as.numeric,Classifier-method}
\alias{as.factor,Classifier-method}
\alias{show,Classifier-method}
\alias{package_version-class}


\title{Class \code{"Classifier"}}
\description{
    A container for the result of a classifier
}

\section{Methods}{
    \describe{
    \item{as.numeric}{: Gets the classifier scores }
    \item{as.factor}{:Gets the classifier classifications}
    }
}
\keyword{classes}
