% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class_ClassifierResults.R
\docType{class}
\name{ClassifierResults}
\alias{ClassifierResults}
\title{An S4 class to store classifier results.}
\description{
This class stores classifier results as obtained after running
the \code{\link{runClassifier}} function.
}
\section{Slots}{

\describe{
\item{\code{classifierParameters}}{An object of class \code{\link{ClassifierParameters}}
in which the applied classifier parameters are stored.}

\item{\code{score}}{A numeric vector of resulting classifier scores}

\item{\code{batchCorrection}}{A character vector indicating wheter batch correction was applied}

\item{\code{weightingType}}{A character string indicating wheter the weighting type was complete
(i.e. no missing data), reweighted (i.e. missing data was handled based on correction
using the covariance structure in the classifiers training data), or reduced (i.e. missing
data but not reweighting the original probeset weighting)}

\item{\code{.geneClassifierVersion}}{An object of class \code{\link{package_version}}}
}}

