% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class_ExpressionSets.R
\docType{methods}
\name{dim,FixedExpressionData-method}
\alias{dim,FixedExpressionData-method}
\title{Dimensions of an Object}
\usage{
\S4method{dim}{FixedExpressionData}(x)
}
\arguments{
\item{x}{an R object, for example a matrix, array or data frame.}
}
\value{
Retrieves the 'dim attribute of the object.  It is 'NULL' or a
    vector of mode 'integer'.
}
\description{
Retrieve the dimension of an object.
}
\examples{
data(exampleMAS5)
myData <- setNormalizationMethod(exampleMAS5, "MAS5.0", targetValue=500)
dim(myData)
dim(myData[1:10,1:3])
}
\seealso{
Other fixed data information extraction functions: \code{\link{[,FixedExpressionData-method}},
    \code{\link{getNormalizationMethod}},
    \code{\link{getTargetValue}}
}

