% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMC92.R
\docType{data}
\name{EMC92_ExpressionSet_training}
\alias{EMC92_ExpressionSet_training}
\title{EMC92 training set}
\description{
An \code{\link{ExpressionSet}}. The data was MAS5.0 normalized to a target value of 500, log2 transformed, mean centered and scaled to unit variance.
}
\value{An object of class \code{ExpressionSet} with 92 rows and 328 columns.}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE19784}
}
\usage{
EMC92_ExpressionSet_training
}

\keyword{datasets}

