\name{eset.mas5}
\alias{eset.mas5}
\docType{data}
\title{
Gene expression example data
}
\description{
This data contains gene expression sample from 25 multiple myeloma patients from the HOVON-65/GMMG-HD4 trial (GEO accession: GSE19784)
The data was mas5 normalized using the affy packages with target value = 500.

}

\references{
\href{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE19784}{GEO accession: GSE19784}
}
\examples{
data(eset.mas5)
}
\keyword{datasets}
