% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class_ExpressionSets.R
\docType{data}
\name{exampleMAS5}
\alias{exampleMAS5}
\title{Example MAS5.0 ExpressionSet}
\description{
An \code{\link{ExpressionSet}}. The data contains a sample of gene expression data from 
patients included in the HOVON65/GMMG-HD4 trial on multiple myeloma. The data was MAS5.0 
normalized to a target value of 500.
}
\value{An object of class \code{ExpressionSet} with 374 rows and 25 columns.}
\usage{
exampleMAS5
}
\keyword{datasets}
