% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneClassifiers.R
\docType{package}
\name{geneClassifiers-package}
\alias{geneClassifiers}
\alias{geneClassifiers-package}
\title{geneClassifiers: Application of gene classifiers}
\description{
This packages aims for easy accessible application of
        classifiers which have been published in literature using an
        ExpressionSet as input.
}
\details{
Combining gene expression profiling data with survival data has led to the 
development of robust outcome predictors (gene classifiers).This package 
provides a method for running gene classifiers generating patient specific 
predictive outcomes. This package is intended to support and enable 
research. To yield stable results, this package requires a dataset of
at least 20 patients.

For detail on how to use this package see the vignette: 
\code{vignette("geneClassifiers")}
}
\references{
Kuiper R, Broyl A, de Knegt YA, Van Vliet MH, Van Beers EH, van der Holt B, el Jarari L, Mulligan G, Gregory W, Morgan G, Goldschmidt H. A gene expression signature for high-risk multiple myeloma. Leukemia. 2012 Nov;26(11):2406.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://doi.org/doi:10.18129/B9.bioc.geneClassifiers}
  \item Report bugs at \url{https://github.com/rkuiper/geneClassifiers/issues}
}

}
\author{
\strong{Maintainer}: R Kuiper \email{r.kuiper.emc@gmail.com} (0000-0002-3703-1762)

}
\keyword{internal}
