% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenericFunctions.R,
%   R/Class_ClassifierResults.R
\docType{methods}
\name{getBatchCorrection}
\alias{getBatchCorrection}
\alias{getBatchCorrection,ClassifierResults-method}
\title{Obtain the batch correction status for a classifier result.}
\usage{
getBatchCorrection(object)

\S4method{getBatchCorrection}{ClassifierResults}(object)
}
\arguments{
\item{object}{An object of class \code{\link{ClassifierResults}}
as returned by \code{\link{runClassifier}}}
}
\value{
TRUE or FALSE
}
\description{
\code{getBatchCorrection} returns TRUE or FALSE
 indicating whether correction was applied
}
\examples{
myData <- setNormalizationMethod(exampleMAS5, "MAS5.0",targetValue=500)
results <- runClassifier('EMC92', myData)
getBatchCorrection( results )
}
\seealso{
Other classifier results: \code{\link{getClassifications}},
  \code{\link{getScores}}, \code{\link{getWeightingType}}
}
\concept{classifier results}
