% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenericFunctions.R,
%   R/Class_TransformationProcess.R, R/Class_ClassifierParameters.R,
%   R/Class_ClassifierResults.R
\docType{methods}
\name{getName}
\alias{getName}
\alias{getName,TransformationProcess-method}
\alias{getName,ClassifierParameters-method}
\alias{getName,ClassifierResults-method}
\title{Obtain object names.}
\usage{
getName(object)

\S4method{getName}{TransformationProcess}(object)

\S4method{getName}{ClassifierParameters}(object)

\S4method{getName}{ClassifierResults}(object)
}
\arguments{
\item{object}{The object to get the name of.}
}
\value{
The return value is a character string
}
\description{
\code{getName} returns the name associated with the requested object.
}
\examples{
aClassifier <- getClassifier("EMC92")
getName( aClassifier )
}
\seealso{
\code{\link{ClassifierParameters}}

\code{\link{ClassifierResults}}
}
