% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenericFunctions.R, R/Class_ClassifierParameters.R
\docType{methods}
\name{getTrainingData}
\alias{getTrainingData}
\alias{getTrainingData,ClassifierParameters-method}
\title{Obtain classifier training data.}
\usage{
getTrainingData(object)

\S4method{getTrainingData}{ClassifierParameters}(object)
}
\arguments{
\item{object}{An object of class \code{\link{ClassifierParameters}}
as returned by \code{\link{getClassifier}}}
}
\value{
An object of class \code{\link{ExpressionSet}}
}
\description{
\code{getTrainingData} returns the training data that was used
for building the classifier.
}
\examples{
aClassifier <- getClassifier("EMC92")
getTrainingData(aClassifier)
}
\seealso{
Other classifier information functions: \code{\link{getCitations}},
  \code{\link{getClassifier}},
  \code{\link{getDecisionBoundaries}},
  \code{\link{getDescription}},
  \code{\link{getEventChain}}, \code{\link{getIntercept}},
  \code{\link{getMeans}},
  \code{\link{getNormalizationMethod}},
  \code{\link{getProbeNames}}, \code{\link{getSds}},
  \code{\link{getWeights}}
}

