% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenericFunctions.R, R/Class_ClassifierResults.R
\docType{methods}
\name{getWeightingType}
\alias{getWeightingType}
\alias{getWeightingType,ClassifierResults-method}
\title{Obtain the weighting type used to obtain a classifier result.}
\usage{
getWeightingType(object)

\S4method{getWeightingType}{ClassifierResults}(object)
}
\arguments{
\item{object}{An object of class \code{\link{ClassifierResults}}
as returned by \code{\link{runClassifier}}}
}
\value{
\code{complete} or \code{reweighted}
}
\description{
\code{getWeightingType} returns weigthing type
}
\examples{
myData <- setNormalizationMethod(exampleMAS5, "MAS5.0",targetValue=500)
results <- runClassifier('EMC92', myData)
getWeightingType( results )
}
\seealso{
Other classifier results: \code{\link{getBatchCorrection}},
  \code{\link{getClassifications}}, \code{\link{getScores}}
}

