% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HM19.R
\docType{data}
\name{HM19_ExpressionSet_training}
\alias{HM19_ExpressionSet_training}
\title{HM19 training set:}
\description{
An \code{\link{ExpressionSet}}. The data was GCRMA normalized.
}
\value{An object of class \code{ExpressionSet} with 19 rows and 206 columns.}
\usage{
HM19_ExpressionSet_training
}

\keyword{datasets}

