% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MILLENNIUM100.R
\docType{data}
\name{MILLENNIUM100_ExpressionSet_training}
\alias{MILLENNIUM100_ExpressionSet_training}
\title{MILLENNIUM100 training set.}
\description{
MILLENNIUM100 training set. An \code{\link{ExpressionSet}}. This is data from AffyPlus2 A and B chips which were
MAS5.0 normalized to a target value of 500. Afterwards they were merged together,log2 
transformed and a scaling toward mean=0, sd=1 per probe set was applied.
}
\value{An object of class \code{ExpressionSet} with 100 rows and 264 columns.}
\usage{
MILLENNIUM100_ExpressionSet_training
}

\keyword{datasets}

