% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MRCIX6.R
\docType{data}
\name{MRCIX6_ExpressionSet_training}
\alias{MRCIX6_ExpressionSet_training}
\title{MRCIX6 training set}
\description{
An \code{\link{ExpressionSet}}. The data was MAS5.0 normalized to a target value of 500 and log2 transformed.
}

\value{An object of class \code{ExpressionSet} with 6 rows and 247 columns.}
\usage{
MRCIX6_ExpressionSet_training
}
\keyword{datasets}

