% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UAMS17.R
\docType{data}
\name{UAMS17_ExpressionSet_training}
\alias{UAMS17_ExpressionSet_training}
\title{UAMS70 training set}
\description{
An \code{\link{ExpressionSet}}. The data was MAS5.0 normalized to a target value of 500, log2 transformed, mean centered and scaled to unit variance.
}
\value{An object of class \code{ExpressionSet} with 17 rows and 345 columns.}
\usage{
UAMS17_ExpressionSet_training
}
\keyword{datasets}

