% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UAMS70.R
\docType{data}
\name{UAMS70_ExpressionSet_training}
\alias{UAMS70_ExpressionSet_training}
\title{UAMS70 training set:}
\description{
An \code{\link{ExpressionSet}}. The data was MAS5.0 normalized to a target value of 500 and log2 transformed.
}
\value{An object of class \code{ExpressionSet} with 70 rows and 345 columns.}
\usage{
UAMS70_ExpressionSet_training
}
\keyword{datasets}

