% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UAMS80.R
\docType{data}
\name{UAMS80_ExpressionSet_training}
\alias{UAMS80_ExpressionSet_training}
\title{UAMS80 training set:}
\description{
An \code{\link{ExpressionSet}}. The data was MAS5.0 normalized to a target value of 500 and log2 transformed.
}
\value{An object of class \code{ExpressionSet} with 80 rows and 380 columns.}
\usage{
UAMS80_ExpressionSet_training
}
\keyword{datasets}

